Updated and Finished Cashiers bug and issues

+++ biosource/application/controls/pos-checkout-delete.php
@@ -8,17 +8,41 @@
 
         require_once('../controls/connection.php');
 
-        $query = "DELETE FROM tbl_checkout WHERE checkout_id = ".$_POST['delid'];
+        $select = "SELECT * FROM tbl_checkout WHERE checkout_id = ".$_POST['delid'];
 
-        $sql = mysqli_query($connection, $query);
+        $query_select = mysqli_query($connection, $select);
 
-        if((int) $sql === 1) {
+        $result = mysqli_fetch_assoc($query_select);
 
-            echo 'success';
+        $update = '';
+
+        if($result['checkout_type'] == 'tbl_product') {
+
+            $update = "UPDATE tbl_product SET product_qtyperpiece = (product_qtyperpiece + ".$result['checkout_qtypiece']."), product_qtyperbox = (product_qtyperbox + ".$result['checkout_qtybox'].") WHERE product_id = ".$result['item_id'];
 
         } else {
 
-            echo 'id-error';
+            $update = "UPDATE tbl_brand SET brand_qtyperpiece = (brand_qtyperpiece + ".$result['checkout_qtypiece']."), brand_qtyperbox = (brand_qtyperbox + ".$result['checkout_qtybox'].") WHERE brand_id = ".$result['item_id'];
+
+        }
+
+        $result_update = mysqli_query($connection, $update);
+
+        if($result_update) {
+
+            $query = "DELETE FROM tbl_checkout WHERE checkout_id = ".$_POST['delid'];
+
+            $sql = mysqli_query($connection, $query);
+
+            if((int) $sql === 1) {
+
+                echo 'success';
+
+            } else {
+
+                echo 'id-error';
+
+            }
 
         }
